mysql: Resume read_package after drop.
This fixes #563

Data read from a mysql stream is ultimitly read by read_package.
If the read package future is droped before it is done, it will
leave the stream in an invalid state. To fix this we store the state of the read_package
future in the stream it self such that read_package just resumens reading the same
package it is called again after beeing dropped